+++ src/components/RickMortyComponent.js
@@ -1,62 +1,52 @@
 import React from 'react';
-import { RickMortyImage } from './RickMortyImage';
-import { RickMortyName } from './RickMortyName';
-import { RickMortyLocation } from './RickMortyLocation';
-import { RickMortyStatus } from './RickMortyStatus';
-import { Container, Row, Col, CardDeck, CardColumns, Card } from 'reactstrap';
+import {
+  Container,
+  Row,
+  Col,
+  CardDeck,
+  CardColumns,
+  Card,
+  CardText
+} from 'reactstrap';
 import styled from 'styled-components';
+import Character from './Character';
 
 function RickMortyComponent(props) {
-    const { rickMortyCharacters } = props;
-    const rickMortyCharactersData = rickMortyCharacters.results;
+  const { rickMortyCharacters } = props;
+  const characterData = rickMortyCharacters.results;
 
-
-
-    return (
-        //   
-            <Container>
-                <RowCustom>
-                
-                            {rickMortyCharactersData.map(rickMortyCharacter => {
-                                return (
-                                    <Col md={6} lg={4} xl={3}>
-                                    <CardCustom key={rickMortyCharacter.id + 'custom'}>
-                                    <CharacterCard key={rickMortyCharacter.id} >
-                                        <RickMortyImage rickMortyCharacter={rickMortyCharacter} key={rickMortyCharacter.id + 'c'} />
-                                        <RickMortyName rickMortyCharacter={rickMortyCharacter} key={rickMortyCharacter.id + 'n'} />
-                                        <RickMortyLocation rickMortyCharacter={rickMortyCharacter} key={rickMortyCharacter.id + 'l'} />
-                                        <RickMortyStatus rickMortyCharacter={rickMortyCharacter} key={rickMortyCharacter.id + 's'} />
-                                    </CharacterCard>
-            
-                                    </CardCustom>
-                                    </Col>
-                                
-                                )
-                            })}
-                              
-                </RowCustom>  
-            </Container>
-      
-    )
-
+  return (
+    //
+    <Container>
+      <RowCustom>
+        {characterData.map(character => {
+          return (
+            <Col md={6} lg={4} xl={3} key={character.id}>
+              <Character character={character} />
+            </Col>
+          );
+        })}
+      </RowCustom>
+    </Container>
+  );
 }
 export default RickMortyComponent;
 
-const CardCustom = styled(Card)`
-	background-color: rgba(226, 106, 153, 0.19);
-	width: 100%;
-    display: block;
-    border: 1px solid red;
-`;
+// const CardCustom = styled(Card)`
+//   background-color: rgba(226, 106, 153, 0.19);
+//   width: 100%;
+//   display: block;
+//   border: 1px solid red;
+// `;
 
 const RowCustom = styled(Row)`
-    /* display: flex;
+  /* display: flex;
     flex-wrap: wrap; */
 `;
 
-const CharacterCard = styled(Card)`
-	background-color: rgba(226, 106, 153, 0.19);
-	width: 100%;
-    display: block;
-    border: 1px solid red;
-`;
+// const CharacterCard = styled(Card)`
+//   background-color: rgba(226, 106, 153, 0.19);
+//   width: 100%;
+//   display: block;
+//   border: 1px solid red;
+// `;
