--- src/components/RickMortyLocation.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export function RickMortyLocation(props) {
-    const { rickMortyCharacter } = props;
-    const location = rickMortyCharacter.location.name;
-    return (
-        <p>{location}</p>
-    )
-}
